refactor des vues
